jumping: return the jumped-from node if there is no node to jump to.
Previously, the last node in the next-chain was returned (ie. the first
one without a next node), but this can lead to infinite loops, which is
irritating and unexpected.

For example, if a snippet is expanded in a $0, its `next` is the
surrounding $0. If it becomes active upon jumping to the end of the
snippet, the next jump will lead back into the snippet, leading to an
infinite loop.