manually adjust insertion-point in snippet, if nearby node is better.
the binary search may lead to a textNode, despite an adjacent insertNode
being a much better target.
This is only a heuristic, and will not lead to perfect results (does
not, for example, detect dynamicNode with only textNodes inside) but
should handle the common case pretty well, and it's no disaster if the
parent is not chosen as desired.