fix: make exitNode behave correctly if there is no real next node.
"real next node" means that there might be a exitNode.next, but the
cursor wouldn't stop there. "real next node" is thus a node that the
cursor can stop at.
The split between insert/exitNode makes this a bit easier, since only
exitNode has to take care here, insertNodes always have at least an
exitNode (0 at the end, or -1 at the beginning of the snippet. This
holds even if the snippet is inserted into a parentNode)